Refactor tsv export
Creating a dict first has the obvious disadvantage that the
original order of the fields is lost and also might cause
unneccessary errors. Export now is done directly from the xml
and it is possible to add the order of fields for each record
in the future if necessary as an additional column.