Change SQLAlchemy session handling
Whoever calls the functions will have to have an SQLAlchemy Session
handy. This should make writing tests easier and is more in line
with how SQLAlchemy is supposed to be used.

db_setup now only contains the table definitions. Anything related
to the connection to the database was moved to db_connect, which should
be used to establish a SQLAlchemy Session based on the env vars.